requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
comm==0.2.3
contourpy==1.3.0
cvzone==1.5.6
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
easyocr==1.7.2
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.19.1
filterpy==1.4.5
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
lap==0.4.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.4
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.13.0
numpy==1.23.4
openalpr==1.1.0
opencv-python==4.8.1.78
opencv-python-headless==4.7.0.72
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.3
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytubefix==9.5.0
pytz==2025.2
PyWavelets==1.6.0
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
requests==2.32.4
scikit-image==0.19.3
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tifffile==2024.8.30
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.0
ultralytics-thop==2.0.15
urllib3==2.5.0
wcwidth==0.2.13
yarl==1.20.1
zipp==3.23.0
